fix: print error location in console
