+++ packages/dbml-cli/src/errors/index.js
@@ -1,0 +1,7 @@
+import DomainError from './domainError';
+import SyntaxError from './syntaxError';
+
+export {
+  DomainError,
+  SyntaxError,
+};
